+++ pylearn2/sandbox/cuda_convnet/conv_util.cuh
@@ -676,7 +676,7 @@ void convLocalPool(NVMatrix& images, NVMatrix& target, int numFilters,
     int numImages = images.getNumCols();
     int imgPixels = images.getNumRows() / numFilters;
     assert(images.getNumRows() == numFilters * imgPixels);
-    int imgSize = int(sqrt(imgPixels));
+    int imgSize = int(sqrt((double)imgPixels));
     assert(imgSize * imgSize == imgPixels);
     
     assert(!images.isTrans());
